+++ src/content-scripts/lib/signal.ts
@@ -9,7 +9,7 @@ export default class Signal {
     // This class can accepts multiple request to indicate busy state. However there can only 1 busy signal presented to user. This is meant to avoid crowded noise to user.
     // Every 'busy' request will have a `busyId`. When no longer `busyId` registered, 'vusy' state will be stopped.
     private busyIds: number[] = [];
-    private nextBusyId = 1;
+    private nextBusyId = 0;
     // Interval used to singal 'busy' state
     private busyIntervalId: ReturnType<typeof setInterval> | null = null;
 
@@ -21,16 +21,17 @@ export default class Signal {
      * @return number The id to signal that this process no longer busy. 
      */
     async busy(): Promise<number> {
+        this.nextBusyId++;
         this.busyIds.push(this.nextBusyId);
         this.evaluteBusyState();
-        return this.nextBusyId++;
+        return this.nextBusyId;
     }
 
     /**
      * Stop busy signal
      */
     async busyStop(busyId: number) {
-        for (let i = 1; i < this.busyIds.length; i++) {
+        for (let i = 0; i < this.busyIds.length; i++) {
             if (this.busyIds[i] === busyId) {
                 this.busyIds.splice(i, 1);
                 break;
@@ -47,9 +48,10 @@ export default class Signal {
         if (this.busyIntervalId === null
             && this.busyIds.length > 0
         ) {
+            this.pl.speak('Loading.');
             this.busyIntervalId = setInterval(() => {
-                this.pl.speak('Loading...');
-            }, 6e3);
+                this.pl.speak('Loading.');
+            }, 5e3);
         } else if (this.busyIntervalId !== null
             && this.busyIds.length === 0
         ) {